Stabilize Debian branch of the role to not pull from unstable
This commit changes the default source for python 3.6 to the least
invasive source repository, which is not Debian unstable, but, at the time of the
commit, testing. To ensure the role doesn't pull testing in
unnecessarily, we use the release name, such that once buster gets
releasted as the stable distribution, the role still works as intended
and will continue doing so at least for the time of buster being the
stable release of Debian GNU/Linux.