Update Debian build to work on latest Stretch (#10)
* Don't use with_items in apt calls anymore

The `apt` module now supports calling it with a list of packages in
`name` and the `with_items` syntax is deprecated.

* Debian: Update Python to version 3.7

The Python version in Buster has changed from 3.6, so we need to
upgrade the references to Python 3.7.

* Add requirements for building wheels and messages

As we are building some dependencies and compling translations, we need
to have gcc, gettext and Python development tools installed.

* Install redis server on Debian too

On Arch Linux, the role will install a Redis server. For consistency,
it should do so on Debian too.